socket.js - Add support for real user ip if behind a proxy
NOTE! security section not tested.  With or without this update when I tic "Use uibuilder's security?" in uibuilder node's property window I get console errors about JWT and a constant reboot loop of node-red. With or without this update

Need someone that understands the security section to test if this is working.